added RoomConnectionContainer
