--- src/components/RTCContainer.js
@@ -1,9 +1,0 @@
-import * as React from 'react';
-
-class RTCContainer extends React.Component {
-    render() {
-        return <></>;
-    }
-}
-
-export default RTCContainer;
