+++ src/config/PlanetImages.js
@@ -1,0 +1,1 @@
+export const PlanetImages = ["#jupiter", "#saturn", "#neptune", "#pluto", "#uranus"];
